[Templates] Update DesktopGL template
