--- ProjectTemplates/DotNetTemplate/MonoGame.Templates.CSharp/content/BlankApp.DesktopGL/.template.config/dotnetcli.host.json
@@ -1,3 +1,0 @@
-{
-
-}
