+++ ProjectTemplates/DotNetTemplate/MonoGame.Templates.CSharp/content/MonoGame.Application.DesktopGL/Program.cs
@@ -1,15 +1,9 @@
 ﻿using System;
 
-namespace BlankApp.Desktop
+namespace MGNamespace
 {
-    /// <summary>
-    /// The main class.
-    /// </summary>
     public static class Program
     {
-        /// <summary>
-        /// The main entry point for the application.
-        /// </summary>
         [STAThread]
         static void Main()
         {
